policy: Split indexer into smaller modules
The `index` submodule is getting pretty unwieldy. It's probably time to
start breaking it up.

split index into smaller files

modularize indexer
