+++ policy-controller/core/src/lib.rs
@@ -5,10 +5,7 @@ pub mod http_route;
 mod identity_match;
 mod network_match;
 
-pub use self::{
-    http_route::InboundHttpRoute, http_route::OutboundHttpRoute, identity_match::IdentityMatch,
-    network_match::NetworkMatch,
-};
+pub use self::{identity_match::IdentityMatch, network_match::NetworkMatch};
 use ahash::AHashMap as HashMap;
 use anyhow::Result;
 use futures::prelude::*;
@@ -32,7 +29,7 @@ pub struct InboundServer {
 
     pub protocol: ProxyProtocol,
     pub authorizations: HashMap<AuthorizationRef, ClientAuthorization>,
-    pub http_routes: HashMap<InboundHttpRouteRef, InboundHttpRoute>,
+    pub http_routes: HashMap<http_route::inbound::HttpRouteRef, http_route::inbound::HttpRoute>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -46,12 +43,6 @@ pub enum AuthorizationRef {
     Default(&'static str),
     ServerAuthorization(String),
     AuthorizationPolicy(String),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum InboundHttpRouteRef {
-    Default(&'static str),
-    Linkerd(String),
 }
 
 /// Describes how a proxy should handle inbound connections.
@@ -109,29 +100,8 @@ pub type OutboundPolicyStream = Pin<Box<dyn Stream<Item = OutboundPolicy> + Send
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct OutboundPolicy {
-    pub http_routes: HashMap<String, OutboundHttpRoute>,
+    pub http_routes: HashMap<String, http_route::outbound::HttpRoute>,
     pub authority: String,
     pub namespace: String,
     pub opaque: bool,
-}
-
-// === impl InboundHttpRouteRef ===
-
-impl Ord for InboundHttpRouteRef {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        match (self, other) {
-            (Self::Default(a), Self::Default(b)) => a.cmp(b),
-            (Self::Linkerd(a), Self::Linkerd(b)) => a.cmp(b),
-            // Route resources are always preferred over default resources, so they should sort
-            // first in a list.
-            (Self::Linkerd(_), Self::Default(_)) => std::cmp::Ordering::Less,
-            (Self::Default(_), Self::Linkerd(_)) => std::cmp::Ordering::Greater,
-        }
-    }
-}
-
-impl PartialOrd for InboundHttpRouteRef {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
 }
